+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v1.1.2
+
+- Added ligature wildcard font, which is used by fonts without specific settings when global ligatures are enabled.
+
+- Fixed the checkbox: prevented multiple or blank selections, and made sure clicking the checkbox changes the theme.
+
 ## v1.1.1
 
 - Fix conflict of random theme changes on a timer when multiple windows are open